Work rest cycle
Still coding some parts but here are the current features this has

- Tabs
- Work and Rest input

current bugs
Visual bugs
- tabs are shown during counter
- when rest is active, the entry keeps the last entered number in place

Will soon add a way to pause the counter and save the current count to a progress bar after fixing a few bugs.

the tabs as of now are empty but will be the next update.